+++ src/main/java/nextstep/courses/domain/session/SessionImageType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import java.util.Arrays;
