+++ src/main/java/nextstep/courses/domain/session/SessionState.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public enum SessionState {
     PREPARING,
