+++ src/test/java/nextstep/courses/domain/session/SessionCapacityTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.exception.CannotIncreaseException;
 import org.junit.jupiter.api.Test;
