+++ src/test/java/nextstep/courses/domain/session/SessionImageTypeTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import org.junit.jupiter.api.Test;
